Fix behavior when --export/-e is omitted
In v2, when --export/-e was omitted, the workflow would not export; this
was made possible by the 'const' parameter to
argparse.ArgumentParser.add_argument. However, Python does not natively
support using `const` with `nargs="*"`, so when we recently switched no
`nargs="*"` in 6013f91, the behavior changed.

Fortunately, per the StackOverflow post I referenced within the code, we
can restore the desired behavior via a custom action on the argument.